tutorboard/requirements.txt
Django==1.6.9
django-debug-toolbar==1.3.0
django-filter==0.10.0
django-haystack==2.4.0
djangorestframework==3.1.3
Markdown==2.6.2
MySQL-python==1.2.5
Pillow==2.5.3
South==1.0.2
sqlparse==0.1.15
Whoosh==2.7.0
